Avoid using getBucket in S3FileSystemProvider
When using S3FileStore::getBucket the AWS user needs the permission 's3:ListAllMyBuckets'. In S3FileSystemProvider::createDirectory it is checked if a bucket exists, but this uses S3FileStore::getBucket thus needing the additional permission, even if this shouldn't be the case.
This commit adds S3FileStore::hasBucket to avoid the mentioned method call in S3FileSystemProvider.